fix: use `mod_interface`
